requirements.txt
django==4.2.2
python-dotenv==1.0.0
psycopg2==2.9.6  # узнать ставить ли binary
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
djoser==2.2.0